correction et ajout des erreurs de frappes sur les migrations

+++ database/migrations/2025_05_02_150855_create_modele_vehicules_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('modele_vehicules', function (Blueprint $table) {
             $table->id();
             $table->string('nom_modele');
-            $table->integer('tarif');
+            $table->float('tarif');
             $table->timestamps();
         });
     }
